Chore: trigger cnpm sync on publish (#333)
* chore: trigger cnpm sync on publish

* chore: bump version

* fix: simple publish script not call cnpm sync script

* fix: do not delete action tokens that are scanned

+++ packages/did-auth/lib/handlers/wallet.js
@@ -198,7 +198,12 @@ module.exports = class WalletHandlers {
         }
 
         onExpire({ token, status: 'expired' });
-        await this.storage.delete(token);
+        const store = await this.storage.read(token);
+
+        // We do not delete tokens that are scanned by wallet since it will cause confusing
+        if (store && store.status !== STATUS_SCANNED) {
+          await this.storage.delete(token);
+        }
         res.status(200).json({ token });
       } catch (err) {
         res.json({ error: err.message });
